BlockChain.Transactions (block_chain v0.1.0)

Documentation for Transactions.

This module is responsible for saving all transactions that are not yet added to the blockchain as blocks.

Functions

add_transaction(wallet, transaction)

@spec add_transaction(
  %BlockChain.Wallet{
    private_key: term(),
    public_key: term(),
    transactions: term()
  },
  %BlockChain.Transaction{amount: term(), from: term(), to: term()}
) :: :ok

Add a new transaction in the blockchain and updates the wallet.

Parameters

wallet: The wallet to save.

transaction: The transaction to save.

Returns :ok if the transaction was saved successfully.

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear_transactions()

@spec clear_transactions() :: :ok

Clear all transactions from the agent.

get_transaction_count()

@spec get_transaction_count() :: integer()

Get the trascation count.

Returns the number of transactions as an integer.

get_transactions()

@spec get_transactions() :: [
  %BlockChain.Transaction{amount: term(), from: term(), to: term()}
]

Get all transactions.

Returns a list of all transactions not yet added to the blockchain.

start_link(opts)

Starts the Transactions Agent.