BlockChain.Wallet (block_chain v0.1.0)

Documentation for Wallet.

A Wallet is a digital wallet that contains a list of transactions and a public and private key.

Functions

add_transaction(wallet, transaction)

@spec add_transaction(
  %BlockChain.Wallet{
    private_key: term(),
    public_key: term(),
    transactions: term()
  },
  %BlockChain.Transaction{amount: term(), from: term(), to: term()}
) :: %BlockChain.Wallet{
  private_key: term(),
  public_key: term(),
  transactions: term()
}

Add a transaction to the wallet, without registering it in the blockchain. Do not call this method directly, use BlockChain.Transaction.new/3 instead.

Parameters

wallet: The wallet to add the transaction to.

transaction: The transaction to add.

balance(wallet)

@spec balance(%BlockChain.Wallet{
  private_key: term(),
  public_key: term(),
  transactions: term()
}) ::
  integer()

Get the balance of the wallet.

Parameters

wallet: The wallet to get the balance of.

new()

@spec new() :: %BlockChain.Wallet{
  private_key: term(),
  public_key: term(),
  transactions: term()
}

Creates a new wallet.